server/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.3.2               # via kombu
asn1crypto==0.24.0        # via cryptography
billiard==3.5.0.5         # via celery
celery==4.2.1             # via girder-worker
certifi==2018.11.29       # via requests
cffi==1.11.5              # via cryptography
chardet==3.0.4            # via requests
click==7.0                # via girder-client
cryptography==2.4.2       # via pyopenssl, requests
decorator==4.3.0          # via networkx
diskcache==3.1.1          # via girder-client
docker-pycreds==0.4.0     # via docker
docker==3.6.0             # via girder-worker
girder-client==2.4.0      # via girder-worker, girder-worker-utils
girder-worker-utils==0.8.4  # via girder-worker
girder-worker==0.5.0
idna==2.8                 # via cryptography, requests
jsonpickle==1.0           # via girder-worker, girder-worker-utils
kombu==4.2.2.post1        # via celery
networkx==1.11            # via girder-worker
pbr==5.1.1                # via stevedore
pycparser==2.19           # via cffi
pymongo==3.7.2            # via girder-worker
pyopenssl==18.0.0         # via requests
pytz==2018.9              # via celery
requests-toolbelt==0.8.0  # via girder-client
requests[security]==2.21.0  # via docker, girder-client, girder-worker, requests-toolbelt
setuptools-scm==3.1.0     # via girder-worker, girder-worker-utils
six==1.12.0               # via cryptography, docker, docker-pycreds, girder-client, girder-worker, girder-worker-utils, pyopenssl, stevedore, websocket-client
stevedore==1.30.0         # via girder-worker
urllib3==1.24.1           # via requests
vine==1.2.0               # via amqp
websocket-client==0.54.0  # via docker
